Refactor experiment archives directory structure
Moved radial and spring-mass experiment files into dedicated subdirectories for better organization. Updated import paths in affected modules to reflect new locations. Also increased the default frequency boost in spherical_wave.py from 0.1 to 1.0.
